Move getUnpaidWorkOptions method to UnpaidWorkUtils
We will be introducing the requirement page to journey for creating a new
appointment.

As a result, moving this method to the utils so it can be reused in both
places.
