+++ server/utils/unpaidWorkUtils.ts
@@ -28,6 +28,28 @@ export default class UnpaidWorkUtils {
     }
   }
 
+  static getUnpaidWorkOptions(unpaidWorkDetails: Array<UnpaidWorkDetailsDto>, selectedOptionValue?: number) {
+    return unpaidWorkDetails.map(detail => {
+      const text = detail.mainOffence.description
+      const value = detail.eventNumber
+
+      const { totalHoursOrdered, eteHoursCredited, eteHoursRemaining } = UnpaidWorkUtils.unpaidWorkHoursDetails(detail)
+
+      const hintHtml = [
+        `Event number: ${detail.eventNumber}`,
+        `Sentence date: ${DateTimeFormats.isoDateToUIDate(detail.sentenceDate)}`,
+        `Total hours ordered: ${totalHoursOrdered}`,
+        `ETE hours credited: ${eteHoursCredited}`,
+        `ETE hours remaining: ${eteHoursRemaining}`,
+        `Status: ${detail.upwStatus}`,
+      ].join('<br>')
+
+      const checked = detail.eventNumber === selectedOptionValue
+
+      return { text, value, hint: { html: hintHtml }, checked }
+    })
+  }
+
   private static totalHoursRemaining(detail: UnpaidWorkDetailsDto) {
     return DateTimeFormats.totalMinutesToHumanReadableHoursAndMinutes(detail.requiredMinutes - detail.completedMinutes)
   }
